chore: move getMappingAnnotaion
